feat(blog): Let users extend/customize elblog routes
The default elblog route (/{article-name}) can now be extended with
user-supplied routes by overriding the elblog-additional-routes
customize variable.

This variable takes the same format as the alist supplied to
elnode-hostpath-dispatcher.

The prefix `/en` has been dropped from the default handler because
that only existed to be compatible with my old blog, which is no
longer required as it can now be handled with a custom legacy route in
my personal blog configuration.