feat(cgit-taz): Use cheddar as the about-filter
This will render about pages using the Comrak renderer defined in
Cheddar.

Note that due to the way its implemented this will have one
interesting behaviour: Markdown files in the tree will *also* be
rendered as HTML.

I will need to see how that works out before deciding whether or not
to disable it.